Implement a badRequests counter
I just tried to use the Hystrix Dashboard and it broke because of a missing `rollingCountBadRequests` property.

This adds that counter, and the required stuff around it.

What I'm unsure about here is how a bad request should be tracked, it's not very well documented in Hystrix itself (...or, I just couldn't find the docs). I found this issue: https://github.com/Netflix/Hystrix/issues/186 that discussed bad requests in relationship to the dash.

BREAKING CHANGE: This is a breaking change for anyone implementing custom collectors as the collector interface has a new function `IncrementBadRequests()`.